requirements.txt
gradio==6.12.0
gradio-imageslider==0.0.20
h5py==3.16.0
matplotlib==3.7.5
numpy==1.24.4
onnx==1.17.0
onnxruntime==1.19.2
open3d==0.19.0
opencv-python==4.12.0.88
Pillow==12.2.0
pysodmetrics==1.6.2
PyYAML==6.0.3
tensorboardX==2.6.5
thop==0.1.1.post2209072238
timm==1.0.26
torch>=1.9.0
torchvision>=0.10.0
tqdm==4.67.1
xformers==0.0.35